Changed int measures to double

+++ src/Actions/Action.cs
@@ -10,7 +10,7 @@ namespace StarshipWanderer.Actions
         public string Name { get; set; }
 
         /// <inheritdoc/>
-        public int Attitude { get; set; }
+        public double Attitude { get; set; }
 
         /// <summary>
         /// Initializes action with a default <see cref="Name"/> based on the class name
